Refactor code style: rename s_class to Class, use native types, block bodies for methods
- Rename s_class to Class in generator and hand-written files (NSString pattern)
- Change int/uint property/method types to nint/nuint with MsgSendPtr/MsgSendNUInt
- Remove int<->nint and uint<->nuint parameter type conversions (MapParamType)
- Remove (nuint) casts from generated parameter passing
- Keep expression body for read-only properties, block body for methods
- Update NSError.Code to return nint, NSArray.ObjectAtIndex param to nint
- Regenerate all bindings

+++ Metal.NET/Foundation/NSArray.cs
@@ -29,7 +29,7 @@ public class NSArray : IDisposable
         return new(value);
     }
 
-    public nint ObjectAtIndex(int index)
+    public nint ObjectAtIndex(nint index)
     {
         return ObjectiveCRuntime.MsgSendPtr(NativePtr, NSArraySelector.ObjectAtIndex, index);
     }
